Add files via upload
These are all of the files that I've either used or tried to use when I was developing this theme. Such as different icons for possible icon designs and the actual things you need to develop this project. It include files and ideas of the designs that could've been added to my theme, If I was smart enough or if it was even possible lol. I hope you can enjoy my custom theme :).